+++ src/ast.ml
@@ -113,7 +113,7 @@ and expr =
   | PipelineDef of pipeline_node list
   | IntentDef of (string * expr) list
 
-and binop = Plus | Minus | Mul | Div | Eq | NEq | Gt | Lt | GtEq | LtEq | And | Or | Pipe
+and binop = Plus | Minus | Mul | Div | Eq | NEq | Gt | Lt | GtEq | LtEq | And | Or | Pipe | MaybePipe
 and unop = Not | Neg
 and comp_clause = CFor of { var : symbol; iter : expr } | CFilter of expr